+++ juniper/src/http/mod.rs
@@ -36,7 +36,7 @@ where
     S: ScalarValue,
 {
     /// Returns the `operation_name` associated with this request.
-    fn operation_name(&self) -> Option<&str> {
+    pub fn operation_name(&self) -> Option<&str> {
         self.operation_name.as_ref().map(|oper_name| &**oper_name)
     }