+++ juniper_rocket/CHANGELOG.md
@@ -1,6 +1,6 @@
 # master
 
-- No changes yet
+- Expose the operation name from `GraphQLRequest`.
 
 # [0.2.0] 2018-12-17