+++ juniper_rocket/CHANGELOG.md
@@ -1,6 +1,6 @@
 # master
 
-- No changes yet
+- Compatibility with the latest `juniper`.
 
 # [0.2.0] 2018-12-17