+++ juniper_rocket/CHANGELOG.md
@@ -1,6 +1,7 @@
 # master
 
 - Expose the operation name from `GraphQLRequest`.
+- Compatibility with the latest `juniper`.
 
 # [0.2.0] 2018-12-17